reorganise js files
