+++ app.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="description" content="QA test page for Atomic Jamboree." />
     <meta name="keywords" content="atomic, jamboree, qa, test" />
-    <link rel="stylesheet" href="./qa-info.css" />
+    <link rel="stylesheet" href="/src/qa-info.css" />
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
       rel="stylesheet"
@@ -16,9 +16,7 @@
     <div id="nav-bar"></div>
     <div id="app"></div>
 
-    <script type="module" src="/navbar.js"></script>
-    <script type="module" src="/main.js"></script>
-    <script type="module" src="./qa-info.js"></script>
+    <script type="module" src="/src/main.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
   </body>
 </html>
