+++ src/components/badgePlacement.js
