+++ src/qa-info.css
