+++ src/shared/initAtomicCommerce.js
