+++ src/shared/initAtomicLoader.js
