+++ src/shared/initAtomicRecommendations.js
