+++ src/shared/initBadgePlacements.js
